+++ src/main/java/org/folio/linked/data/mapper/dto/monograph/instance/sub/CopyrightEventMapperUnit.java
@@ -28,10 +28,10 @@ public class CopyrightEventMapperUnit implements InstanceSubResourceMapperUnit {
   private final HashService hashService;
 
   @Override
-  public <P> P toDto(Resource source, P parentDto, Resource parentResource) {
+  public <P> P toDto(Resource resourceToConvert, P parentDto, ResourceMappingContext context) {
     if (parentDto instanceof InstanceResponse instance) {
-      var copyrightEvent = coreMapper.toDtoWithEdges(source, CopyrightEventResponse.class, false);
-      copyrightEvent.setId(String.valueOf(source.getId()));
+      var copyrightEvent = coreMapper.toDtoWithEdges(resourceToConvert, CopyrightEventResponse.class, false);
+      copyrightEvent.setId(String.valueOf(resourceToConvert.getId()));
       instance.addCopyrightItem(copyrightEvent);
     }
     return parentDto;