+++ src/main/java/org/folio/linked/data/mapper/dto/monograph/work/sub/DissertationMapperUnit.java
@@ -32,9 +32,9 @@ public class DissertationMapperUnit implements WorkSubResourceMapperUnit {
   private final HashService hashService;
 
   @Override
-  public <P> P toDto(Resource source, P parentDto, Resource parentResource) {
-    var dissertation = coreMapper.toDtoWithEdges(source, DissertationResponse.class, false);
-    dissertation.setId(String.valueOf(source.getId()));
+  public <P> P toDto(Resource resourceToConvert, P parentDto, ResourceMappingContext context) {
+    var dissertation = coreMapper.toDtoWithEdges(resourceToConvert, DissertationResponse.class, false);
+    dissertation.setId(String.valueOf(resourceToConvert.getId()));
     if (parentDto instanceof WorkResponse work) {
       work.addDissertationItem(dissertation);
     }