+++ NEWS.md
@@ -21,6 +21,7 @@
 - *BREAKING* Update okapiInterfaces for Split Search & Browse APIs. [MODLD-778](https://folio-org.atlassian.net/browse/MODLD-778)
 - Update API to support 'Series Statement' (MARC 490) [MODLD-777](https://folio-org.atlassian.net/browse/MODLD-777)
 - Add `on delete cascade` constraint to folio_metadata table. Remove method `ResourceGraphService -> breakEdgesAndDelete` [MODLD-780](https://folio-org.atlassian.net/browse/MODLD-780)
+- Update API to support type of Language (Marc 041) [MODLD-362](https://folio-org.atlassian.net/browse/MODLD-362)
 
 ## 1.0.4 (04-24-2025)
 - Work Edit form - Instance read-only section: "Notes about the instance" data is not shown [MODLD-716](https://folio-org.atlassian.net/browse/MODLD-716)